Improve location display in `TrashablesView`.
The location's `NameOrUniqueName` actually seems unique enough to use as a key, e.g. it's different for different sheds on the farm, but the `DisplayName` is often not helpful at all, e.g. it displays "<X> Farm" whether outside or in the farmhouse.

Fix is to lift the location-formatting code from the unfinished SC mod, which handles the whole hierarchy.