De-unique-ify mine floor location names.
Every floor has a different name, which makes the per-location trash filter fairly useless in these areas with the previous implementation. Switches to a unique "key" which is based on the name, but eliminates the floor suffix and pulls in the context (e.g. "Desert") which should make different mine "areas" still unique.

Mine floors also don't have meaningful names and the `GetPathText` extension was returning an empty string, which made the menu a bit confusing; now falls back to the key if no name can be constructed. (There might be a better way to get a name like "Skull Cavern"; I couldn't find it).